+++ src/Detail.js
@@ -15,13 +15,13 @@ import CreateButton from './CreateButton'
 import {
   EditSaveButton,
   EditCancelButton,
-  InlineInput,
   isFieldEditing,
   getFieldEditData,
   InlineEditButton,
   FileDelete,
   TableEditButton,
-  getFieldErrorEdit
+  getFieldErrorEdit,
+  EditInput
 } from './Edit'
 import { Popover, PopoverContent } from './Popover'
 import getDisplayValue from './utils/getDisplayValue'
@@ -41,12 +41,12 @@ const LabelInfoPopover = ({ LabelInfoComponent, fieldLabel }) => (
   />
 )
 
-export const DefaultDetailLabel = ({ schema, modelName, fieldName, data, customProps }) => {
+export const DefaultDetailLabel = ({ schema, modelName, fieldName, node, customProps }) => {
   const LabelInfoComponent = R.path(['components', 'labelInfo'], getField(schema, modelName, fieldName))
   if (skipOverride(LabelInfoComponent)) {
     return null
   }
-  const fieldLabel = getFieldLabel({ schema, modelName, fieldName, data, customProps })
+  const fieldLabel = getFieldLabel({ schema, modelName, fieldName, node, customProps })
   if (LabelInfoComponent) {
     return <LabelInfoPopover {...{ LabelInfoComponent, fieldLabel }} />
   }
@@ -99,19 +99,21 @@ export const DefaultDetailAttribute = ({
       <React.Fragment>
         <dt className='col-sm-3 text-sm-right'>
           {
-            skipOverride(LabelOverride) ? null : <DetailLabel {...{ schema, modelName, fieldName, data: node, customProps }} />
+            skipOverride(LabelOverride) ? null : <DetailLabel {...{ schema, modelName, fieldName, node, customProps }} />
           }
         </dt>
         <dd className='col-sm-9'>
-          <InlineInput {...{
-            schema,
-            modelName,
-            fieldName,
-            node,
-            editData: fieldEditData,
-            error,
-            selectOptions
-          }} />
+          <div className='detail-edit d-inline-block pull-left'>
+            <EditInput {...{
+              schema,
+              modelName,
+              fieldName,
+              node,
+              editData: fieldEditData,
+              error,
+              selectOptions
+            }} />
+          </div>
           <div className='inline-btn-group'>
             <EditSaveButton {...{
               onClick: (fieldType === 'file')
@@ -148,7 +150,7 @@ export const DefaultDetailAttribute = ({
       <React.Fragment>
         <dt className='col-sm-3 text-sm-right'>
           {
-            skipOverride(LabelOverride) ? null : <DetailLabel {...{ schema, modelName, fieldName, data: node, customProps }} />
+            skipOverride(LabelOverride) ? null : <DetailLabel {...{ schema, modelName, fieldName, node, customProps }} />
           }
         </dt>
         <dd className='col-sm-9'>
@@ -205,7 +207,7 @@ export const DefaultDetailO2MTableTitle = ({ schema, modelName, fieldName, targe
 
   return (
     <DefaultDetailTableTitleWrapper>
-      <DefaultDetailLabel {...{ schema, modelName, fieldName, data: node, customProps }} />
+      <DefaultDetailLabel {...{ schema, modelName, fieldName, node, customProps }} />
       { creatable && <DetailCreateButton {...{
         schema,
         modelName,
@@ -233,7 +235,7 @@ const DefaultDetailM2MTableTitle = ({
   const editable = isFieldEditable({ schema, modelName, fieldName, node, user, customProps })
   return (
     <div style={{ marginBottom: '10px' }}>
-      <h4 className='d-inline'>{getFieldLabel({ schema, modelName, fieldName, data: node, customProps })}</h4>
+      <h4 className='d-inline'>{getFieldLabel({ schema, modelName, fieldName, node, customProps })}</h4>
       {editable && <div className='pl-2 d-inline'>
         <TableEditButton {...{
           schema,
@@ -265,7 +267,7 @@ const DefaultDetailM2MFieldLabel = ({
   const required = R.prop('required', getField(schema, modelName, fieldName))
   const Label = () => (
     <div style={{ marginBottom: '10px' }}>
-      <h4 className='d-inline'>{getFieldLabel({ schema, modelName, fieldName, data: node, customProps })}</h4>
+      <h4 className='d-inline'>{getFieldLabel({ schema, modelName, fieldName, node, customProps })}</h4>
       { required && ' *'}
       { creatable && <DetailCreateButton {...{
         schema,
@@ -324,8 +326,7 @@ export const DefaultDetailTable = ({
           targetModelName,
           user,
           customProps
-        }}>{getFieldLabel({schema, modelName, fieldName, data: node, customProps})}
-        </DetailLabel>
+        }} />
         }
         { skipOverride(ValueOverride) ? null : <DetailValue
           key={`Table-${id}-${targetModelName}-${fieldName}`}
@@ -481,7 +482,7 @@ export const partitionDetailFields = ({ schema, modelName, node, include = null,
 }
 
 const DefaultDetailPageTitle = ({ schema, modelName, node, modalData, user, customProps }) => {
-  const model = getModelLabel({ schema, modelName, data: node, user, customProps })
+  const model = getModelLabel({ schema, modelName, node, user, customProps })
   const label = getDisplayValue({ schema, modelName, node, customProps })
   const actions = getActions(schema, modelName)
   const onDelete = R.path(['delete', 'onDetailDeleteFromDetailPage'], actions)