+++ CHANGES.md
@@ -13,6 +13,8 @@
 -   Add error to Int field overflow #57
 -   Fix filter and sort button header position #52
 -   Expose onFileChange method for custom File inputs #43
+-   Del InlineInput component #81
+-   Fix node and data props #70
 
 ## Version 1.2.0